Move inner processing result class to separate file
