+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/EnvelopeEventProcessor.java
@@ -15,6 +15,7 @@ import uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.IMessageOpe
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.IProcessedEnvelopeNotifier;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.NotificationSendingException;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.exceptions.InvalidMessageException;
+import uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.handler.MessageProcessingResult;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.handler.MessageProcessingResultType;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.model.Envelope;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
@@ -191,19 +192,5 @@ public class EnvelopeEventProcessor implements IMessageHandler {
             : baseMessage;
 
         log.error(fullMessage, exception);
-    }
-
-    class MessageProcessingResult {
-        public final MessageProcessingResultType resultType;
-        public final Exception exception;
-
-        public MessageProcessingResult(MessageProcessingResultType resultType) {
-            this(resultType, null);
-        }
-
-        public MessageProcessingResult(MessageProcessingResultType resultType, Exception exception) {
-            this.resultType = resultType;
-            this.exception = exception;
-        }
     }
 }
